src/modules/auth-module/requirements.txt
Flask>=2.0
PyJWT>=2.0
Werkzeug>=2.0
